Allow cudaMalloc(ptr, 0)
This makes the behaviour more like malloc.
This is beneficial because a program may excpect a subroutine to request for a certain amount of memory, and allocate the requested amount.
They don't check whether this amount is == 0 and thus may cause crashes if the requested amount is equal to 0.
The OpenCL spec states that clCreateBuffer calls with size 0 must error, but malloc(0) must not error.